Trait pallet_evm::Config

pub trait Config: Config + Config {
    type FeeCalculator: FeeCalculator;
    type GasToWeight: GasToWeight;
    type CallOrigin: EnsureAddressOrigin<Self::Origin>;
    type WithdrawOrigin: EnsureAddressOrigin<Self::Origin, Success = Self::AccountId>;
    type AddressMapping: AddressMapping<Self::AccountId>;
    type Currency: Currency<Self::AccountId>;
    type Event: From<Event<Self>> + Into<Self::Event>;
    type Precompiles: Precompiles;
    type ChainId: Get<u64>;
    type Runner: Runner<Self>;
    fn config() -> &'static EvmConfig { ... }
}

EVM module trait

Associated Types

type FeeCalculator: FeeCalculator

Calculator for current gas price.

type GasToWeight: GasToWeight

Maps Ethereum gas to Substrate weight.

type CallOrigin: EnsureAddressOrigin<Self::Origin>

Allow the origin to call on behalf of given address.

type WithdrawOrigin: EnsureAddressOrigin<Self::Origin, Success = Self::AccountId>

Allow the origin to withdraw on behalf of given address.

type AddressMapping: AddressMapping<Self::AccountId>

Mapping from address to account id.

type Currency: Currency<Self::AccountId>

Currency type for withdraw and balance storage.

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

type Precompiles: Precompiles

Precompiles associated with this EVM engine.

type ChainId: Get<u64>

Chain ID of EVM.

type Runner: Runner<Self>

EVM execution runner.

Loading content...
Provided methods

fn config() -> &'static EvmConfig

EVM config used in the module.

Loading content...