Trait pallet_evm::ConvertAccountId

pub trait ConvertAccountId<A> {
    fn convert_account_id(account_id: &A) -> H160;
}

Trait for converting account ids of balances module into H160 for EVM module.

Accounts and contracts of this module are stored in its own storage, in an Ethereum-compatible format. In order to communicate with the rest of Substrate module, we require an one-to-one mapping of Substrate account to Ethereum address.

Required methods

fn convert_account_id(account_id: &A) -> H160

Given a Substrate address, return the corresponding Ethereum address.

Loading content...