pallet_evm

Trait OnChargeEVMTransaction

pub trait OnChargeEVMTransaction<T: Config> {
    type LiquidityInfo: Default;

    // Required methods
    fn withdraw_fee(
        who: &H160,
        fee: EvmBalance,
    ) -> Result<Self::LiquidityInfo, Error<T>>;
    fn correct_and_deposit_fee(
        who: &H160,
        corrected_fee: EvmBalance,
        base_fee: EvmBalance,
        already_withdrawn: Self::LiquidityInfo,
    ) -> Self::LiquidityInfo;
    fn pay_priority_fee(tip: Self::LiquidityInfo);
}

Handle withdrawing, refunding and depositing of transaction fees. Similar to OnChargeTransaction of pallet_transaction_payment

Required Associated Types

type LiquidityInfo: Default

Required Methods

fn withdraw_fee( who: &H160, fee: EvmBalance, ) -> Result<Self::LiquidityInfo, Error<T>>

Before the transaction is executed the payment of the transaction fees need to be secured.

fn correct_and_deposit_fee( who: &H160, corrected_fee: EvmBalance, base_fee: EvmBalance, already_withdrawn: Self::LiquidityInfo, ) -> Self::LiquidityInfo

After the transaction was executed the actual fee can be calculated. This function should refund any overpaid fees and optionally deposit the corrected amount, and handles the base fee rationing using the provided OnUnbalanced implementation. Returns the NegativeImbalance - if any - produced by the priority fee.

fn pay_priority_fee(tip: Self::LiquidityInfo)

Introduced in EIP1559 to handle the priority tip.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.