Trait pallet_evm::PrecompileHandle

pub trait PrecompileHandle {
    fn call(
        &mut self,
        to: H160,
        transfer: Option<Transfer>,
        input: Vec<u8, Global>,
        gas_limit: Option<u64>,
        is_static: bool,
        context: &Context
    ) -> (ExitReason, Vec<u8, Global>);
    fn record_cost(&mut self, cost: u64) -> Result<(), ExitError>;
    fn remaining_gas(&self) -> u64;
    fn log(
        &mut self,
        address: H160,
        topics: Vec<H256, Global>,
        data: Vec<u8, Global>
    ) -> Result<(), ExitError>;
    fn code_address(&self) -> H160;
    fn input(&self) -> &[u8];
    fn context(&self) -> &Context;
    fn is_static(&self) -> bool;
    fn gas_limit(&self) -> Option<u64>;
}

Handle provided to a precompile to interact with the EVM.

Required Methods

fn call(
&mut self,
to: H160,
transfer: Option<Transfer>,
input: Vec<u8, Global>,
gas_limit: Option<u64>,
is_static: bool,
context: &Context
) -> (ExitReason, Vec<u8, Global>)

Perform subcall in provided context. Precompile specifies in which context the subcall is executed.

fn record_cost(&mut self, cost: u64) -> Result<(), ExitError>

Record cost to the Runtime gasometer.

fn remaining_gas(&self) -> u64

Retreive the remaining gas.

fn log(
&mut self,
address: H160,
topics: Vec<H256, Global>,
data: Vec<u8, Global>
) -> Result<(), ExitError>

Record a log.

fn code_address(&self) -> H160

Retreive the code address (what is the address of the precompile being called).

fn input(&self) -> &[u8]

Retreive the input data the precompile is called with.

fn context(&self) -> &Context

Retreive the context in which the precompile is executed.

fn is_static(&self) -> bool

Is the precompile call is done statically.

fn gas_limit(&self) -> Option<u64>

Retreive the gas limit of this call.