Trait pallet_evm::PrecompileSet

pub trait PrecompileSet {
    fn execute(
        &self,
        handle: &mut impl PrecompileHandle
    ) -> Option<Result<PrecompileOutput, PrecompileFailure>>;
    fn is_precompile(&self, address: H160) -> bool;
}

A set of precompiles. Checks of the provided address being in the precompile set should be as cheap as possible since it may be called often.

Required Methods

fn execute(
&self,
handle: &mut impl PrecompileHandle
) -> Option<Result<PrecompileOutput, PrecompileFailure>>

Tries to execute a precompile in the precompile set. If the provided address is not a precompile, returns None.

fn is_precompile(&self, address: H160) -> bool

Check if the given address is a precompile. Should only be called to perform the check while not executing the precompile afterward, since execute already performs a check internally.