Crate pallet_evm

EVM execution module for Substrate

Re-exports

pub use crate::precompiles::Precompile;

pub use crate::precompiles::Precompiles;

Modules

precompiles
Builtin precompiles.

Structs

Account
Ethereum account nonce, balance and code. Used by storage.

Backend
Substrate backend for EVM.

EnsureAddressNever
Ensure that the origin never happens.

EnsureAddressRoot
Ensure that the origin is root.

EnsureAddressSame
Ensure that the EVM address is the same as the Substrate address. This only works if the account ID is H160.

EnsureAddressTruncated
Ensure that the address is truncated hash of the origin. Only works if the account id is AccountId32.

GenesisAccount
Account definition used for genesis block construction.

GenesisConfig
Genesis config for the module, allow to build genesis storage.

HashedAddressMapping
Hashed address mapping.

IdentityAddressMapping
Identity address mapping.

Log
Ethereum log. Used for deposit_event.

Module

SystemChainId
Substrate system chain ID.

Vicinity
External input from the transaction.

Enums

Call
Dispatchable calls.

Error

ExitError
Exit error reason.

ExitFatal
Exit fatal reason.

ExitReason
Exit reason.

ExitRevert
Exit revert reason.

ExitSucceed
Exit succeed reason.

RawEvent
Events for this module.

Traits

AddressMapping

EnsureAddressOrigin

FeeCalculator
Trait that outputs the current transaction gas price.

Trait
EVM module trait

Type Definitions

BalanceOf
Type alias for currency balance.

Event
RawEvent specialized for the configuration Trait