Trait pallet_evm::pallet::Config

pub trait Config: Config + Config {

    type FeeCalculator: FeeCalculator;
    type GasWeightMapping: GasWeightMapping;
    type WeightPerGas: Get<u64>;
    type BlockHashMapping: BlockHashMapping;
    type CallOrigin: EnsureAddressOrigin<Self::Origin>;
    type WithdrawOrigin: EnsureAddressOrigin<Self::Origin, Success = Self::AccountId>;
    type AddressMapping: AddressMapping<Self::AccountId>;
    type Currency: Currency<Self::AccountId> + Inspect<Self::AccountId>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type PrecompilesType: PrecompileSet;
    type PrecompilesValue: Get<Self::PrecompilesType>;
    type ChainId: Get<u64>;
    type BlockGasLimit: Get<U256>;
    type Runner: Runner<Self>;
    type OnChargeTransaction: OnChargeEVMTransaction<Self>;
    type FindAuthor: FindAuthor<H160>;

    fn config() -> &'static EvmConfig { ... }

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type FeeCalculator: FeeCalculator

Calculator for current gas price.

type GasWeightMapping: GasWeightMapping

Maps Ethereum gas to Substrate weight.

type WeightPerGas: Get<u64>

Weight corresponding to a gas unit.

type BlockHashMapping: BlockHashMapping

Block number to block hash.

type CallOrigin: EnsureAddressOrigin<Self::Origin>

Allow the origin to call on behalf of given address.

type WithdrawOrigin: EnsureAddressOrigin<Self::Origin, Success = Self::AccountId>

Allow the origin to withdraw on behalf of given address.

type AddressMapping: AddressMapping<Self::AccountId>

Mapping from address to account id.

type Currency: Currency<Self::AccountId> + Inspect<Self::AccountId>

Currency type for withdraw and balance storage.

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type PrecompilesType: PrecompileSet

Precompiles associated with this EVM engine.

type PrecompilesValue: Get<Self::PrecompilesType>

type ChainId: Get<u64>

Chain ID of EVM.

type BlockGasLimit: Get<U256>

The block gas limit. Can be a simple constant, or an adjustment algorithm in another pallet.

type Runner: Runner<Self>

EVM execution runner.

type OnChargeTransaction: OnChargeEVMTransaction<Self>

To handle fee deduction for EVM transactions. An example is this pallet being used by pallet_ethereum where the chain implementing pallet_ethereum should be able to configure what happens to the fees Similar to OnChargeTransaction of pallet_transaction_payment

type FindAuthor: FindAuthor<H160>

Find author for the current block.

Provided Methods

fn config() -> &'static EvmConfig

EVM config used in the module.